[ALPS05652867] mml: refine tile code
- Reduce cross reference from platform related code
- Simplify parameter types
- Slim more unused code

MTK-Commit-Id: 7a46996b9f113fd576bb38be8e1727afa4e6b623

CR-Id: ALPS05652867
Feature: [Module]MML driver

+++ drivers/gpu/drm/mediatek/mml/mtk-mml-tile.h
@@ -9,7 +9,7 @@
 
 #include "mtk-mml-core.h"
 
-s32 calc_tile(struct mml_task *task, u8 pipe_idx);
+s32 calc_tile(struct mml_task *task, u32 pipe_idx);
 
 void destroy_tile_output(struct mml_tile_output *output);
